requirements.txt
fastapi
pydantic
sqlalchemy
uvicorn
psycopg2-binary
asyncpg
celery[pytest]
aio-pika
python-dotenv
redis
python-jose
python-multipart
bcrypt
broadcaster